requirements.txt
duckdb
snowflake-connector-python[pandas]